fix: 프로필 입력 경고와 feature client 구조를 정리
